请求加入权重

+++ braavos/src/binance/bn_models.rs
@@ -335,9 +335,7 @@ pub struct CommandInfo {
     pub base: BinanceBase,
     pub path: BinancePath,
     pub has_security: bool,
-    pub weight: i16
-    // pub security: Option<SecurityInfo>,
-    // pub client: &'a reqwest::Client,
+    pub weight: u32
 }
 
 pub struct TimeStampRequest {
